webui: extract MountModal, use ButtonGroup, fix CSS
- Extract MountModal into its own file (mountModal.jsx)
- Wrap mount and copy buttons in a Bootstrap ButtonGroup
  with outline-secondary variant
- Fix .object-viewer-buttons CSS to use direct child selector
  (> .btn) so ButtonGroup internal margins aren't overridden
- Remove unused lodash/noop import
- Update screenshots
