+++ webui/src/lib/components/repository/tree.jsx
@@ -22,13 +22,12 @@ import Table from 'react-bootstrap/Table';
 import Card from 'react-bootstrap/Card';
 import OverlayTrigger from 'react-bootstrap/OverlayTrigger';
 import Button from 'react-bootstrap/Button';
+import ButtonGroup from 'react-bootstrap/ButtonGroup';
 import Container from 'react-bootstrap/Container';
 import Row from 'react-bootstrap/Row';
 import Col from 'react-bootstrap/Col';
 import Dropdown from 'react-bootstrap/Dropdown';
 import Modal from 'react-bootstrap/Modal';
-import Form from 'react-bootstrap/Form';
-import InputGroup from 'react-bootstrap/InputGroup';
 import { FaDownload } from 'react-icons/fa';
 
 import { commits, linkToPath, objects } from '../../api';
@@ -38,8 +37,8 @@ import { Link } from '../nav';
 import { RefTypeBranch, RefTypeCommit } from '../../../constants';
 import { ClipboardButton, copyTextToClipboard, AlertError, Loading, TooltipButton } from '../controls';
 import { useAPI } from '../../hooks/api';
-import noop from 'lodash/noop';
 import { CommitInfoCard } from './commits';
+import MountModal from './mountModal';
 
 export const humanSize = (bytes) => {
     if (!bytes) return '0.0 B';
@@ -459,45 +458,6 @@ const PathLink = ({ repoId, reference, path, children, presign = false, as = nul
     return React.createElement(as, { href: link, download: name }, children);
 };
 
-const MountModal = ({ show, onHide, repo, reference, path }) => {
-    const mountCommand = `everest mount lakefs://${repo.id}/${reference.id}/${path}`;
-    return (
-        <Modal show={show} onHide={onHide} size={'lg'}>
-            <Modal.Header closeButton>
-                <Modal.Title>Mount Directory</Modal.Title>
-            </Modal.Header>
-            <Modal.Body>
-                <p>
-                    Mount this path as a local directory using{' '}
-                    <a href="https://docs.lakefs.io/reference/mount.html" target="_blank" rel="noopener noreferrer">
-                        Everest
-                    </a>
-                    :
-                </p>
-                <InputGroup>
-                    <Form.Control readOnly value={mountCommand} className="font-monospace" disabled />
-                </InputGroup>
-                <p className="text-muted mt-3">
-                    Mount is only available with{' '}
-                    <a href="https://lakefs.io/enterprise/" target="_blank" rel="noopener noreferrer">
-                        lakeFS Enterprise
-                    </a>{' '}
-                    or{' '}
-                    <a href="https://lakefs.cloud/register" target="_blank" rel="noopener noreferrer">
-                        lakeFS Cloud
-                    </a>
-                    .
-                </p>
-            </Modal.Body>
-            <Modal.Footer>
-                <Button variant="secondary" onClick={onHide}>
-                    Close
-                </Button>
-            </Modal.Footer>
-        </Modal>
-    );
-};
-
 const EntryRow = ({ config, repo, reference, path, entry, onDelete, showActions }) => {
     let rowClass = 'change-entry-row ';
     switch (entry.diff_type) {
@@ -880,28 +840,28 @@ export const URINavigator = ({
                 {breadcrumbItems.length > 0 && <CollapsibleBreadcrumb items={breadcrumbItems} minVisibleItems={2} />}
             </div>
             <div className="object-viewer-buttons" style={{ flexShrink: 0 }}>
-                {hasMountButton && !isPathToFile && (
-                    <TooltipButton
-                        variant="link"
-                        size="sm"
-                        tooltip="Mount directory"
-                        className="me-1"
-                        onClick={() => setShowMountModal(true)}
-                    >
-                        <TerminalIcon />
-                    </TooltipButton>
-                )}
-                {hasCopyButton && (
-                    <ClipboardButton
-                        text={`lakefs://${repo.id}/${reference.id}/${path}`}
-                        variant="link"
-                        size="sm"
-                        onSuccess={noop}
-                        onError={noop}
-                        className={'me-1'}
-                        tooltip={'copy URI to clipboard'}
-                    />
-                )}
+                <ButtonGroup size="sm" className="me-1">
+                    {hasMountButton && !isPathToFile && (
+                        <TooltipButton
+                            variant="outline-secondary"
+                            size="sm"
+                            tooltip="Mount directory"
+                            onClick={() => setShowMountModal(true)}
+                        >
+                            <TerminalIcon />
+                        </TooltipButton>
+                    )}
+                    {hasCopyButton && (
+                        <ClipboardButton
+                            text={`lakefs://${repo.id}/${reference.id}/${path}`}
+                            variant="outline-secondary"
+                            size="sm"
+                            onSuccess={() => {}}
+                            onError={() => {}}
+                            tooltip={'copy URI to clipboard'}
+                        />
+                    )}
+                </ButtonGroup>
                 {downloadUrl && (
                     <a
                         href={downloadUrl}
