+++ webui/src/styles/objects/objects.css
@@ -201,11 +201,11 @@ tr:hover .row-hover .btn-outline-danger {
     align-items: center;
 }
 
-.object-viewer-buttons .btn {
+.object-viewer-buttons > .btn {
     margin-right: var(--spacing-md);
 }
 
-.object-viewer-buttons .btn:last-child {
+.object-viewer-buttons > .btn:last-child {
     margin-right: 0;
 }
